chore: polish code by review comments
